+++ src/common/secfrac_digits/owned.rs
@@ -224,6 +224,21 @@ impl TryFrom<Vec<u8>> for SecfracDigitsString {
     }
 }
 
+impl TryFrom<String> for SecfracDigitsString {
+    type Error = ConversionError<String>;
+
+    #[inline]
+    fn try_from(v: String) -> Result<Self, Self::Error> {
+        match validate_bytes(v.as_bytes()) {
+            Ok(_) => Ok(unsafe {
+                // This is safe because the value is successfully validated.
+                Self::from_bytes_maybe_unchecked(v.into_bytes())
+            }),
+            Err(e) => Err(ConversionError::new(v, e)),
+        }
+    }
+}
+
 impl fmt::Display for SecfracDigitsString {
     #[inline]
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {